static extension : Text Transforms
